refactor: 상수와 타입 분리
