+++ src/components/Card/CardFormSection/CardFormSection.tsx
@@ -1,5 +1,6 @@
-import { CARD_FORM_TYPE, CardFormType } from "../../../constants/constants";
+import { CARD_FORM_TYPE } from "../../../constants/constants";
 import { useCardValidation } from "../../../hooks/useCardValidation";
+import { CardFormType } from "../../../types/types";
 import Description from "../../Common/Description/Description";
 import Error from "../../Common/Error/Error";
 import Subtitle from "../../Common/Subtitle/Subtitle";
